scripts/requirements.txt
# VC Intelligence Scraper Requirements
requests>=2.31.0
pandas>=2.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
